Main

Documentation

data Manager

Constructors

Manager

Fields

state :: State

pid :: Int

step :: Int

before :: Integer

delta :: Integer

delay :: Integer

loadManager :: Manager

O sitio para mudar o mapa é na tarefa 2 | alterando a variavel startingState e tambem a variavel ghostInsideHouseId

updateControlledPlayer

Arguments

:: Key
Tecla pressionada

-> Manager
Manager antes da alteração

-> Manager
Manager depois da alteração

Função que muda a orientação do player dependendo da tecla que premir.

Funções de dar print à interface no ecrã

updateScreen

Arguments

:: Window
Janela que contem o ecrã que vai ser atualizado

-> ColorID
ID da cor usada por defeito quando não existe cor para uma string

-> Manager
Manager com a informação a usar na interface

-> Curses ()
Resultado do Curses

Função que atualiza o ecrã com a a interface do jogo. | PS: esta função foi alterada para poder integrar as cores do Ncurses

drawPlayerStats

Arguments

:: Int
Tamanho do maze (Variavel usada para saber onde começar as infos)

-> State
State com a informção a ser escrita no ecrã

-> ColorID
ID da cor usada por defeito quando não existe cor para uma string

-> Window
Janela que contem o ecrã que vai ser atualizado

-> Curses ()
Resultado do Curses

Funcao que escreve os player stats no ecra. Parte por baixo do Maze.

drawColorInstructions

Arguments

:: Window
Janela que contem o ecrã que vai ser atualizado

-> ColorInstructions
Instruções com as cores e as respetivas string a serem imprimidas no ecrã

-> Coords
Acumulador de coordenadas

-> Curses Coords
Resultado do Curses (acumulador de coordanadas)

Função responsável por desenhar o labirinto com cores. | Percorre uma lista de ColorInstructions pregerada para desenhar cada string com a cor ddesejada.

drawColorInstruct

Arguments

:: Window
Janela que contem o ecrã que vai ser atualizado

-> ColorInstruction
Instrução unica as cores e as respetivas string correspondentes a uma linha a serem imprimidas no ecrã

-> Coords
Acumulador de coordenadas

-> Curses Coords
Resultado do Curses (acumulador de coordanadas)

Função responsável por desenhar uma unica linha do labirinto com cores.

uLostScreen

Arguments

:: Window
Janela que contem o ecrã que vai ser atualizado

-> ColorID
ID da cor usada por defeito quando não existe cor para uma string

-> Manager
Manager com a informação a usar na interface

-> Curses ()
Resultado do Curses

Função usada quando o pacman perde as vidas todas. Corresponde a um ecrã de "perdeste". | Sustitui a função updateScreen a partir do momento que o pacman morre

Funções de gestão de tempo do jogo

currentTime :: IO Integer

Função que devolve o tempo atual em integer

updateTime

Arguments

:: Integer
Tempo atual

-> Manager
Manager antes da alteração

-> Manager
Manager depois da alteração

Função que faz update ao delta atual do manager quando o delta não é maior que o delay

resetTimer

Arguments

:: Integer
Tempo atual

-> Manager
Manager antes da alteração

-> Manager
Manager depois da alteração

Função que reinicia o contador do delta a zero

nextFrame

Arguments

:: Integer
Tempo atual

-> Manager
Manager antes da alteração

-> Manager
Manager depois da alteração

Função que chama a passTime para as jogadas serem executadas e para o tempo passar no state | Tambem chama a resetTimer para o contador do delta voltar a zero

Funções principais

loop

Arguments

:: Window
Window do jogo

-> Manager
Manager com a informação toda do jogo

-> Curses ()
Resultado do Curses

Loop principal do jogo. | Função que está constantemente a ser invocada para o jogo se manter aberto.

main :: IO ()