Tarefa1

Um Maze

sampleMaze :: Maze

Exemplo

Função principal, junção de todas as auxiliares para gerar um Maze

generateMaze

Arguments

:: Int
comprimento dos corredores do maze

-> Int
altura do maze

-> Int
seed que gera um maze aleatório

-> Maze
maze resultante

Nota: Se o comprimento for menor que 15 ou a altura menor que 10 devolve um erro e não gera um Maze

generateRandom :: Int -> Int -> [Int]

Dada uma seed devolve a lista de n inteiros aleatoriamente gerados

randomNrs :: Int -> Int

Dada uma seed devolve um inteiro aleatóriamente gerado

subList :: Int -> [a] -> [[a]]

Converte uma lista numa lista de listas de tamanho n

convertMaze :: [[Int]] -> Maze

Converte uma lista de listas de inteiros num Maze

convertCorridor :: [Int] -> Corridor

Converte uma lista de inteiros num Corridor usando a função convertPiece

convertPiece :: Int -> Piece

Converte um inteiro numa Piece

n == 3 = Food Big

0 n < 70 <= Food Little

70 <= 99 <= Wall

Funções relativas a adicionar as paredes do maze

addWalls :: Maze -> Int -> Maze

Adiciona as paredes ao Maze

verticalWalls :: Corridor -> Corridor

Adiciona paredes aos limites do(s) corredor(es)

Função relativas a adiconar os túneis

openTunels :: Maze -> Maze

Substituiu os limites do(s) corredor(es) do meio do maze por Emptys (túneis)

Funções relativas à adição da casa dos fantasmas no maze

buildHouse :: Maze -> Maze

Substitui as peças do meio do labirinto pela casa dos fantasmas conforme a altura do maze seja par ou ímpar

takeBegCorr

Arguments

:: Int
número de peças do inicio do corredor a separar

-> Int
índice do corredor

-> Maze
maze onde se encontra o corredor

-> Corridor
parte do corredor inicial onde se encontra a casa e separado desta

Põe aparte o início do corredor

takeEndCorr

Arguments

:: Int
indíce de onde acaba a casa e começa a parte final do corredor

-> Int
indíce do corredor no maze

-> Maze
maze onde se encontra o corredor

-> Corridor
parte final do corredor separado da casa

Põe aparte o fim do corredor