Tarefa2

Função referente à primeira fase da função play e as auxiliares a esta

startingState :: State

ghostInsideHouseId :: Int

houseCoords :: Coords

play :: Play -> State -> State

Primeira fase da função play (correspondente ao player ter ou não a orientação da próxima jogada)

Nota: Se o id não corresponder a nenhum jogador ou o id corresponder a um fantasma devolve o estado original

elemID

Arguments

:: Int
o ID que queremos ver se existe na lista de todos os players

-> [Player]
A lista de todos os players

-> Bool
Resultado de percorrer a lista e verificar se existe

Verifica se o id recebido na função play corresponde a algum jogador da lista

selectPlayer

Arguments

:: Int
ID correspondente ao player selecionado

-> [Player]
Lista de todos os players

-> [Player]
Lista de players com o player selecionado pelo ID à cabeça da lista

Põe o player com id correspondente no inicio da lista de players Nota: Nesta função se o ID não existir na lista gera um resultado infinito, mas como se aplica primeiro a elemID (definida acima) na função play esta só se realiza se o ID existir e por isso não vai returnar um resultado infinito

isPacOrNot :: Player -> Bool

Verifica se o jodador é ou não o Pacman

compareOrientation

Arguments

:: Orientation
A próxima orientação

-> Player
O orientação atual do player

-> Bool
Resultado de verificar se são diferentes

Verifica se a orientação atual é diferente à da próxima jogada

changeOrientation

Arguments

:: Orientation
Orientação a atribuir

-> [Player]
Lista de players com o player a atribuir a orientação à sua cabeça

-> [Player]
Resultado

Muda a orientação do player

Segunda fase da função play e as auxiliares a esta

playPacGhosts :: State -> State

Segunda fase da função play, em que se verifica se na próxima posição existe fantasmas ou se o Pacman entra num túnel e aplica-se o respetivo efeito

nextPosition

Arguments

:: Orientation
A orientação para onde se desloca o player

-> Coords
As coordenadas atuais do player

-> Coords
As coordenadas resultantes de se movimentar na orientação para onde se desloca

Devolve as coordenadas em função da orientação em que se desloca

Funções relacionadas aos túneis

isTunnel :: Maze -> Coords -> Bool

Verifica se as coordenadas correspondem a um túnel

tpTunnel

Arguments

:: Maze
Maze onde os jogadores estão inseridos

-> Coords
coordenada onde o player se encontra antes de passar o túnel

-> Coords
coordenada resultante de passar o túnel

Altera as coordenadas do pacman conforme o lado em que entra no túnel

Funções relacionadas aos fantasmas

isPacDeadAfter

Arguments

:: Player
O jogador escolhido para realizar a jogada

-> [Player]
Lista dos restantes jogadores

-> Bool
Resultado

Verifica se existe um Ghost ,na próxima posição do Pacman, que pode matar o pacman

doesPacEatAfter

Arguments

:: Player
O jogador escolhido para realizar a jogada

-> [Player]
Lista dos restantes jogadores

-> Bool
Resultado

Mesma coisa que isPacDeadAfter mas verifica se existe ghost | ,na próxima posição do Pacman, que possa ser comido

returnGhostsHome

Arguments

:: Maze
O maze onde os jogadores estão inseridos (necessário para calcular as coordenadas da casa)

-> Player
O player que realiza a jogada (necessário para comparar a sua coord com os restasntes players)

-> [Player]
Lista dos restantes jogadores

-> [Player]
Lista dos restantes jogadores após alterar o estado dos fantasmas que foram comidos

Muda o estado de todos os fantasmas que o Pacman tenha comido numa jogada

eatDeadGhosts

Arguments

:: Player
O pacman escolhido para realizar a jogada

-> [Player]
A lista dos restantes players

-> Int
Os pontos resultantes de o pacman comer (ou não) algum fantasma

Devolve a quantidade de pontos correspondentes à quantidade de fantasmas que o Pacman irá comer na próxima jogada

Funções relacionadas à ultima fase da função play

playPacFood :: State -> State

Fase final da função play, em que o pacman muda de coordenadas (ou não) e se atribui os respetivos efeitos da próxima peça

isPiece

Arguments

:: Maze
O maze onde se encontra todas as peças

-> Piece
A peça que queremos verificar se existe

-> Coords
As coordenadas onde estamos a verificar se existe a peça escolhida

-> Bool
Resultado

Verifica se as coordenadas correspondem a uma certa peça

changeGhostsToMega

Arguments

:: [Player]
Lista de todos os jogadores exceto o escolhido para a jogada

-> [Player]
Resultado de mudar todos os fantasmas após o Pacman comer a Big Food

Aplica a todos os fantasmas o efeito correspondente ao Pacman estar em modo Mega

turnBack :: Player -> Orientation

exeGhostPlay :: State -> State

ghostMovement :: State -> State

ghostHitsPacman :: Player -> Player