Tarefa4

Description

Introdução

A Tarefa 4 consistiu em atualizar um estado de jogo, que até então, tarefa 2, era atualizado por via de executar a função play, com base no efeito da passagem de um instante de tempo e que atualizasse não um, mas todos os jogadores e o mapa. = Objetivos Os principais objetivos foram criar as funções que se relacionavam diretamente com o tempo, ou seja, atualizar o número de iterações de acordo com a velocidade do jogador, atualizar o pacman de acordo com o controlo do seu movimento com as setas e atualizar os fantasmas de acordo com o modo do pacman e com o step atual do jogo, tentando sempre fazer com que se parecesse ao jogo original. = Discussão e conclusão Conseguimos atingir todos os objetivos

Documentation

defaultDelayTime :: Integer

switchToScatterTimmings :: [Int]

Steps quando os fantasmas entram em modo Scatter

switchToChaseTimmings :: [Int]

Steps quando os fantasmas entram em modo Chase

passTime

Arguments

:: Int
Step atual

-> State
Estado atual

-> State
Resultado do estado após realizar plays para todos os jogadores

Função objetivo que permite fazer atualizações ao estado com o efeito da passagem de uma intância de tempo

runPlayList :: Int -> [Play] -> State -> State

Executa todas as plays que estão guardadas num lista

processTimings :: State -> Int -> [Play] -> [Play]

Função que dependendo do step, da sua paridade e da velocidade verifica se o Player realiza, ou não, | uma jogada e devolve a lista de plays possíveis

switchGhostsMode :: State -> Int -> State

Substitui o modo do fantasma de acordo com o seu step | Permite que haja diferentes fases de jogo num só em intervalos de tempo diferentes

switchGhostsModeAux :: [Player] -> GhostMode -> [Player]

Auxiliar da switchGhostsMode e que percorre a lista substituindo a cada elemento o seu modo atual por um novo

changeTimeMega :: Int -> State -> State

Função que altera o tempo Mega do pacman em cada jogada e aplica aos fantasmas o resultado de estar ou ainda estar em mega

ghostsNormal :: Int -> [Player] -> [Player]

Função que atribui um modo ao fantasma, Scatter ou Chase, dependendo do step atual do jogo

checkCurrentMode

Arguments

:: Int

-> [Int]
Steps de transicao para chase

-> [Int]
Steps de transicao para scatter

-> GhostMode
Modo para qual o fantasma deve mudar no fim do modo mega do pacman

Verfica o modo ao qual um fantasma frightened tem que regressar no fim do modo mega do pacman | para tal verifica qual o momento de transicao mais pequeno que e menor que o step