Tarefa5

Description

Introdução

A Tarefa 5 consiste em fazer com que os fantasmas persigam o pacman ou fujam dele, quando está em modo Mega, ou seja criar os movimentos deles. = Objetivos Para nós os objetivos foram: * Criar bots para os fantasmas em que cada um tivesse uma estratégia diferente de perseguir o pacman, assim como no jogo original. * Criar 4 tipos de modos para os fantasmas: * o modo Frightened , em que o objetivo dos fantasmas é fugir do pacman estando “mortos”; * o modo Chase, em que cada um dos 4 fantasma persegue-o de acordo com a sua determinada estratégia; * o Modo Scatter, em que cada um dos quatro fantasmas move-se para cada um dos diferentes cantos e circula à volta do obstáculo mais próximo; e por fim o modo InHouse em que o objetivo é fazer com que o fantasma saia da casa. * Os diferentes fantasmas, aos quais nomeamos Blinky, Pinky, Inky e Clyde, teriam que ter um perfil diferente em Chase mode, ou seja, a Blinky perseguiria o pacman de acordo com as suas coordenadas, a Pinky perseguiria o pacman de acordo com 4a coordenada à frente do pacman, a Inky perseguiria a coordenada do objetivo da Blinky multiplicada por 2 e a Clyde perseguiria as coordenadas do pacman se tivesse a uma distância menor ou igual a 8 caso contrário reagiria como se estivesse em Scatter mode. = Discussão e conclusão Apesar de termos atingido todos os objetivos achamos que os bots estão num nível de dificuldade elevada para o jogador

Documentation

ghostPlay :: State -> [Play]

Função objetivo desta tarefa que ao receber o estado atual do jogo devolve a lista de melhores plays | para cada fantasma de modo a reagir ao pacman

decideGhostsActions :: [Player] -> State -> [Play]

Devolve a lista plays que os fantasmas irão realizar de acordo com o seu modo atual

removePacFromList :: [Player] -> [GhoState]

Remove o pacman da lista de players

getGhostMode :: Player -> GhostMode

Recebe o modo do fantasma

chaseMode :: State -> Int -> Play

Função que devolve a play que permite aos fantasmas quando estão vivos | e em modo chase seguirem o pacman, cada um com uma estratégia diferente

profileBlinky :: State -> Int -> Play

Perfil do fantasma, Blinky, com perfil 1, que realiza plays em função do objetivo ser | a coordenada do pacman

profilePinky :: State -> Int -> Play

Perfil do fantasma, Pinky, com perfil 2, que realiza plays em função do seu objetivo ser | a quarta coordenada à frente do pacman de acordo com a orientação do mesmo

profileInky :: State -> Int -> Play

Perfil do fantasma, Inky, com perfil 3, que realiza plays em função do seu objetivo ser | a coordenada que corresponde ao ponto do dobro da distância entre o pacman | e do fantasma, Blinky

profileClyde :: State -> Int -> Play

Perfil do fantasma, Clyde, com perfil 4, que realiza plays consoante a distância | ao pacman, se for maior que 8 tem o mesmo objetivo do fantasma Blinky caso contrário | tem como objetivo o mesmo que teria no modo Scatter

getPlay :: Player -> Maze -> Coords -> Play

De acordo com um jogador, um maze e um objetivo devolve a melhor play | para atingir o objetivo pretendido

getBestPlay

Arguments

:: [(Orientation, Coords)]
lista da orientação para a próxima coordenada em par

-> Coords
coordenada do objetivo (para onde o player pretende ir)

-> Orientation
melhor orientação para atingir o objetivo

Devolve a orientação da próxima melhor coordenada para o player se mover | de acordo com a menor distância da mesma até ao objetivo

distanceOnPlain :: (Coords, Coords) -> Float

Calcula a distância no plano de uma coordenada a outra

getPlaysList :: Maze -> Player -> [(Orientation, Coords)]

Devolve a lista de jogadas possíveis do player sem nunca voltar para trás ou embater numa parede

getPlaysList1 :: Maze -> (Orientation, Coords) -> [(Orientation, Coords)]

Mesma funcionalidade da getPlaysList mas usada na Tarefa6

parseWallPlays :: Maze -> [(Orientation, Coords)] -> [(Orientation, Coords)]

getPossiblePlays

Arguments

:: Player
Player em questão

-> [(Orientation, Coords)]
Lista de todos os pares orientação da próx. jogada e próx coordenada

Função que calcula todas as jogadas possíveis de o jogador realizar sem | nunca voltar para trás no Maze

getPossiblePlays1 :: (Orientation, Coords) -> [(Orientation, Coords)]

Função com a mesma funcionalidade da getPossiblePlays e que será usada na | tarefa 6

translateCoordsByOr :: (Orientation, Coords) -> Int -> Coords

Função que dado um inteiro e uma orientação altera a coordenada | Útil para definir o objetivo de certos perfis dos Fantasmas

getPointOfDoubleDistance :: Coords -> Coords -> Coords

Dadas duas coordenadas devolve a respetiva coordenada ao ponto do dobro da distância | Função usada para calcular o objetivo Inky, que depende das coords do pac e da Blinky

getPacOrAndCoor :: Player -> (Orientation, Coords)

Dado um player recebe um par orientação e coordenadas

scatterMode :: State -> Int -> Play

Dado um state e um perfil de um Fantasma define a coord objetivo quando | se encontra em modo Scatter ou seja cada Fantasma tem correspondência com um canto do maze

frightenedMode :: State -> Int -> Play

Função que devolve uma play aleatorizada e que é executada | quando os fantasmas estão em modo Frightened

inHouseMode :: State -> Int -> Play

Função que permite aos fantasmas que estão na casa sairem e | que é utilizada quando existirem fantasmas em modo InHouse

getRandomIndex :: Int -> Int -> Int