Tarefa6

Description

Introdução

A Tarefa 6 consiste em fazer com que o pacman conlua o jogo realizando e avaliando jogadas. = Objetivos Para nós os objetivos e a estratégia adotada para fazer um bot que movimenta o pacman sozinho foram avaliar as peças de acordo com uma tabela de pontuação, em que uma comida vale mais e um fantasma vale menos, por exemplo. Também uma das estratégias foi criar uma Rose Tree em que cada MappedPlay, possível jogada, estaria associada a uma layer ou camada e que a essa mappedplay estaria associada uma lista de mappedplays, possíveis jogadas após a realização da mesma, para essa e associadas também a uma layer mais pequena, quanto mais baixa fosse a layer menos importaria ou influenciaria a tomada de decisão do bot de acordo com o sistema de pontuação. = Discussão e conclusão Apesar de termos conseguido a implementação de todos os objetivos não conseguimos fazer com que o pacman finalizasse o jogo pois haveria um momento em que entraria em um loop à volta de um obstáculo do mapa, concluimos que tal coisa poderia ser resultante ou do método inefeciente que escolhemos para relizar a tarefa ou da performance do haskell, pois neste último caso tivemos que usar apenas 10 layers (fez com que o bot tivesse uma percentagem de pouco acerto nas jogadas que decidia) de jogadas para o bot avaliar pois mais que isso fazia com que o jogo se tornasse muito lento por estar a calcular um número elevado de possibilidades.

Documentation

bot

Arguments

:: Int
Id do pacman

-> State
State atual do jogo

-> Maybe Play
Play que o pacman irá realizar após avaliação de várias outras jogadas

Função objetivo que devolve a melhor play para alcançar o objetivo de finalizar o jogo

numberOfLayers :: Int

Número de jogadas à frente que o pacman avalia

functionUnion :: State -> Play

Função que junta todas as auxiliares e devolve a play com a melhor pontuação

getMappedPlayList :: [(Orientation, Coords)] -> Maze -> Int -> [MappedPlay]

Função que devolve uma lista de mappedplay

getMappedPlay

Arguments

:: (Orientation, Coords)
Par da próx orientação e próx coordenada

-> Maze
Maze

-> Int
Número da layer

-> MappedPlay
Mappedplay resultante

Função que devolve uma mappedplay

getRouteScore

Arguments

:: MappedPlay
mappedPlay a avaliar

-> ([Double] -> Double)
função que recebe uma lista de pontuações e devolve uma especifica, irá ser a de maxima ou minima pontuação

-> Double
pontuação de uma mappedplay

Função que devolve a pontuação de uma determinada mappedplay | Quanto maior a pontuação, melhor o caminho para o bot

evaluatePiece :: Piece -> Orientation -> Double

Função que atribui uma determinada pontuação a cada peça | Sistema útil para determinar as peças com maior importância para a decisão de movimento do bot

minimumsFst :: [(Double, MappedPlay)] -> [(Double, MappedPlay)]

Devolve as mappedPlays com a menor pontuação

maximumsFst :: [(Double, MappedPlay)] -> [(Double, MappedPlay)]

Devolve as mappedPlays com maior pontuação

isPieceGhost :: Piece -> Bool

Verifica se a peça é um fantasmas

isPieceFood :: Piece -> Bool

Verifica se a peça é uma comida

getPieceFromMaze :: Coords -> Maze -> Piece

Devolve a peça com determinadas coordenadas de um maze