Types

Data Types do jogo

data State

Constructors

State

Fields

maze :: Maze

playersState :: [Player]

level :: Int

type Maze = [Corridor]

type Corridor = [Piece]

data Piece

Constructors

Food FoodType

PacPlayer Player

Empty

Wall

data Player

Constructors

Pacman PacState

Ghost GhoState

data Orientation

Constructors

L

R

U

D

Null

data PacState

Constructors

PacState

Fields

pacState :: PlayerState

timeMega :: Double

openClosed :: Mouth

pacmanMode :: PacMode

data GhoState

Constructors

GhoState

Fields

ghostState :: PlayerState

ghostMode :: GhostMode

type Coords = (Int, Int)

type PlayerState = (Int, Coords, Double, Orientation, Int, Int)

data Mouth

Constructors

Open

Closed

data PacMode

Constructors

Dying

Mega

Normal

data GhostMode

Atribuimos novos modos aos fantasmas para enriquecer o jogo e para a realização da tarefa5

Constructors

Scatter
Modo em que o fantasma está vivo e não segue o pacman, indo para um canto do maze

Chase
Modo que representa que o fantasma está vivo e que o seu objetivo é seguir o pacman

Frightened
Modo que representa que o fantasma está morto, pacman modo mega, e que não segue o pacman

InHouse
Modo que indica que o fantasma está na casa e que o seu objetivo é sair da mesma

data FoodType

Constructors

Big

Little

data CustomColor

Um novo data type que nos permitirá pôr o maze colorido

Constructors

Blue

Green

Purple

Red

Yellow

None

data Play

Constructors

Move Int Orientation

Data Types da Tarefa 6

data MappedPlay

Um novo data type que definimos para a realização da tarefa6 e que | consiste em uma Rose Tree que permite mapear todas as próximas jogadas do pacman

Constructors

MappedPlay

Fields

layer :: Int

nextOr :: Orientation

nextCoords :: Coords

nextPiece :: Piece

nextPlays :: [MappedPlay]

Data Types da Tarefa 3

type Instructions = [Instruction]

data Instruction

Constructors

Instruct [(Int, Piece)]

Repeat Int

type ColorInstructions = [ColorInstruction]

Representação de um maze com colorinstruction

type ColorInstruction = [(String, CustomColor)]

Representação de um corridor com uma lista de strings e cores

Instâncias da classe show dos Data Types

instStateShowNew :: State -> String

Função usado para representar em forma de string um state sem representar o maze

coloredCursesString :: CustomColor -> Curses ColorID

Função que gera os colorID's da curses para cada CustomColor

placePlayersOnMap :: [Player] -> Maze -> Maze

printMaze :: Maze -> String

printPlayerStats :: Player -> String

printPlayerStatsDebug :: Player -> String

Função de debug que dá print de todas as caracteríscas de um player

Funções de acesso e alteração ao player

getPlayerID :: Player -> Int

getPlayerByID :: State -> Int -> Player

getPlayerPoints :: Player -> Int

setPlayerCoords :: Player -> Coords -> Player

getPlayerLives :: Player -> Int

setPlayerLives :: Player -> Int -> Player

setPlayerMode :: Player -> Either PacMode GhostMode -> Player

getPlayerSpeed :: Player -> Double

setPlayerSpeed :: Player -> Double -> Player

getPieceOrientation :: Piece -> Orientation

getPacMode :: Player -> PacMode

getPacmanMode :: Player -> PacMode

getPlayerState :: Player -> PlayerState

getPlayerOrientation :: Player -> Orientation

getPacTimeMega :: Player -> Double

setPacMegaTime :: Player -> Double -> Player

replaceElemInMaze :: Coords -> Piece -> Maze -> Maze

replaceNElem :: Int -> a -> [a] -> [a]

getPlayerCoords :: Player -> Coords

flipPacMouth :: Player -> Player

Outras funções

elemAt :: [a] -> Int -> a

elemAtSpec :: [Orientation] -> Int -> Orientation